DAOs implementadas
* interfaces dao

* feat: QuadroDAO

* Lista DAO implementation

* WIP UsuarioDAO

* feat: Usuario DAO implementation

* feat: DAO Projeto, equipe, notificacao, usuario

* feat: DAO UsuarioNotificao

+++ src/main/java/edu/curso/goodooit/domain/repository/IConviteDAO.java
@@ -8,11 +8,11 @@ import java.util.List;
 public interface IConviteDAO {
     public List<Convite> buscarTodosConvites() throws SQLException;
 
-    public Convite ConviteIdRemetente(Long idRemetente) throws SQLException;
+    public Convite ConviteIdRemetente(Integer idRemetente) throws SQLException;
 
-    public Convite buscarConviteIdDestinatario(Long idDestinatario) throws SQLException;
+    public Convite buscarConviteIdDestinatario(Integer idDestinatario) throws SQLException;
 
-    public Convite buscarConviteIdProjeto(Long idProjeto) throws SQLException;
+    public Convite buscarConviteIdProjeto(Integer idProjeto) throws SQLException;
 
     public Convite registrarConvite(Convite convite) throws SQLException;
