+++ src/main/java/edu/curso/goodooit/domain/repository/IStatusTarefaDAO.java
@@ -8,11 +8,11 @@ import java.util.List;
 public interface IStatusTarefaDAO {
     public List<StatusTarefa> buscarTodosStatusTarefa() throws SQLException;
 
-    public StatusTarefa buscarStatusTarefaId(Long id) throws SQLException;
+    public StatusTarefa buscarStatusTarefaId(Integer id) throws SQLException;
 
     public StatusTarefa registrarStatusTarefa(StatusTarefa statusTarefa) throws SQLException;
 
-    public List<StatusTarefa> buscarStatusTarefaProjetoId(Long idProjeto) throws SQLException;
+    public List<StatusTarefa> buscarStatusTarefaProjetoId(Integer idProjeto) throws SQLException;
 
     public void excluirStatusTarefa(StatusTarefa statusTarefa) throws SQLException;
